Endringer (refactor) etter code review
