+++ src/main/kotlin/no/nav/arbeidsgiver/iatjenester/metrikker/datakatalog/metrikker/VirksomhetMetadata.kt
@@ -1,24 +1,11 @@
 package no.nav.arbeidsgiver.iatjenester.metrikker.datakatalog.metrikker
 
+import no.nav.arbeidsgiver.iatjenester.metrikker.datakatalog.Næringskode5Siffer
 import no.nav.arbeidsgiver.iatjenester.metrikker.tilgangskontroll.Orgnr
+import java.time.LocalDate
+import java.time.Month
+import java.time.temporal.ChronoUnit
 
-data class Næringskode5Siffer(var kode: String?, val beskrivelse: String) {
-    init {
-        if (kode.isNullOrBlank()) {
-            throw IllegalArgumentException("Kode for næring kan IKKE være null")
-        }
-        val næringskodeUtenPunktum: String = kode!!.replace(".", "")
-
-        if (erGyldigNæringskode(næringskodeUtenPunktum)) {
-            this.kode = næringskodeUtenPunktum
-        } else {
-            throw IllegalArgumentException("Kode for næring skal være 5 siffer")
-        }
-    }
-
-    private fun erGyldigNæringskode(verdi: String): Boolean = verdi.matches(Regex("^[0-9]{5}$"))
-
-}
 
 interface Virksomhet {
     val orgnr: Orgnr
@@ -44,9 +31,32 @@ data class Underenhet(
     val antallAnsatte: Int
 ) : Virksomhet
 
-data class InstitusjonellSektorkode(
-    val kode: String,
-    val beskrivelse: String
-)
+data class InstitusjonellSektorkode(val kode: String, val beskrivelse: String)
+
+data class Fylke(val nummer: String, val navn: String)
+data class Kommune(val nummer: String, val navn: String)
 
 
+fun mapTilFylke(kommune: Kommune): Fylke {
+
+    val utleddFylkenummer: String =
+        if (kommune.nummer.isNullOrBlank() || kommune.nummer.length != 5)
+            "IKKE_GYLDIG_KOMMUNENUMMER"
+        else
+            kommune.nummer.substring(0, 2)
+
+    return when (utleddFylkenummer) {
+        "42" -> Fylke("42", "Agder")
+        "34" -> Fylke("34", "Innlandet")
+        "15" -> Fylke("15", "Møre og Romsdal")
+        "18" -> Fylke("18", "Nordland")
+        "03" -> Fylke("03", "Oslo")
+        "11" -> Fylke("11", "Rogaland")
+        "54" -> Fylke("54", "Troms og Finnmark")
+        "50" -> Fylke("50", "Trøndelag")
+        "38" -> Fylke("38", "Vestfold og Telemark")
+        "46" -> Fylke("46", "Vestland")
+        "30" -> Fylke("30", "Viken")
+        else -> Fylke("UKJENT", "UKJENT")
+    }
+}
