+++ src/main/kotlin/no/nav/arbeidsgiver/iatjenester/metrikker/enhetsregisteret/EnhetsregisteretService.kt
@@ -8,24 +8,24 @@ import org.springframework.stereotype.Component
 import java.security.InvalidParameterException
 
 @Component
-class EnhetsregisteretOpplysningerService(private val enhetsregisteretClient: EnhetsregisteretClient) {
+class EnhetsregisteretService(private val enhetsregisteretClient: EnhetsregisteretClient) {
 
-    fun hentOpplysningerForUnderenhet(orgnrUnderenhet: Orgnr): Either<EnhetsregisteretException, Underenhet> {
+    fun hentUnderenhet(orgnrUnderenhet: Orgnr): Either<EnhetsregisteretException, Underenhet> {
 
         return try {
-            val underenhet: Underenhet = enhetsregisteretClient.hentInformasjonOmUnderenhet(orgnrUnderenhet)
+            val underenhet: Underenhet = enhetsregisteretClient.hentUnderenhet(orgnrUnderenhet)
             Either.Right(underenhet)
         } catch (e: Exception) {
             Either.Left(EnhetsregisteretException(e.message, e))
         }
     }
 
-    fun hentOpplysningerForOverordnetEnhet(orgnrOverordnetEnhet: Orgnr?):
+    fun hentOverordnetEnhet(orgnrOverordnetEnhet: Orgnr?):
             Either<EnhetsregisteretException, OverordnetEnhet> {
         orgnrOverordnetEnhet?.let {
             return try {
                 val overordnetEnhet: OverordnetEnhet =
-                    enhetsregisteretClient.hentInformasjonOmEnhet(orgnrOverordnetEnhet)
+                    enhetsregisteretClient.hentOverordnetEnhet(orgnrOverordnetEnhet)
                 Either.Right(overordnetEnhet)
             } catch (e: Exception) {
                 Either.Left(EnhetsregisteretException(e.message, e))
